renamed "REST action" to "HTTP action", #76
